[TASK] Substitute deprecated strftime() function with date()
Since PHP 8.1 the strftime() function is deprecated. The function
is substituted with the recommended date() function. The default
$labelFormat argument is adjusted as well to reflect this change.
As the $labelFormat argument in the method signature is not used
elsewhere in the extension this should be sufficient and not
considered breaking.

See: https://wiki.php.net/rfc/deprecations_php_8_1#strftime_and_gmstrftime